Restructure and fix the setup
