+++ setup.py
@@ -5,20 +5,21 @@ def readme():
         return f.read()
 
 setup(name='osmtogtfs',
-      version='0.1.1',
+      version='0.1.4',
       description='Extracts partial GTFS feed from OSM data.',
       long_description=readme(),
       author='Mehdi Sadeghi',
       author_email='mehdi@mehdix.org',
       url='https://github.com/hiposfer/osmtogtfs',
-      py_modules=['osmtogtfs', 'gtfs_writer', 'osm_processor'],
+      packages=['_osmtogtfs'],
+      py_modules=['osmtogtfs'],
       entry_points='''
         [console_scripts]
         osmtogtfs=osmtogtfs:cli
       ''',
       license='MIT',
       platforms='any',
-      install_requires=['osmium', 'timezonefinder', 'pytest', 'click'],
+      install_requires=['osmium', 'timezonefinder', 'click'],
       keywords=['osm', 'gtfs'],
       classifiers=['Development Status :: 4 - Beta',
                    'Environment :: Console',
